crypto: Fix build on AMD64
- isrcry_sha1_final() was attempting to use STORE32H() to store a
64-bit quantity, when what it actually wanted was the lower 32
bits of that quantity.  This caused build failures on AMD64.  Fix this.

+++ crypto/sha1.c
@@ -102,8 +102,10 @@ exported void isrcry_sha1_final(struct isrcry_sha1_ctx *ctx,
 	/* This is slightly inefficient, as the numbers are converted to
 	   big-endian format, and will be converted back by the compression
 	   function. It's probably not worth the effort to fix this. */
-	STORE32H(bitcount >> 32, ctx->block + (SHA1_DATA_SIZE - 8));
-	STORE32H(bitcount, ctx->block + (SHA1_DATA_SIZE - 4));
+	STORE32H((uint32_t)(bitcount >> 32),
+				ctx->block + (SHA1_DATA_SIZE - 8));
+	STORE32H((uint32_t) bitcount,
+				ctx->block + (SHA1_DATA_SIZE - 4));
 	
 	_isrcry_sha1_compress(ctx->digest, ctx->block);
